Added ability to purchase tickets for each item
Added the ability for users to purchase raffle tickets for items.
Currently it can only buy 1 ticket at a time, and does not take the users credits into account or modify the value. These are easy, fast changes.

In the future, users will not be able to buy tickets for their own listings, but for testing purposes we will leave this in for now